Refactor test namespace to denote local event aggregator tests
