+++ src/Lite.EventAggregator.Tests/Lite.EventAggregator.Tests.csproj
@@ -13,7 +13,4 @@
     <Using Include="Microsoft.VisualStudio.TestTools.UnitTesting" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="EventAggr\" />
-  </ItemGroup>
 </Project>
